requirements.txt
# openai>=1.40.0,<2
requests
python-dotenv>=1.0.0,<2
pypdf>=4.0.0,<6
pymupdf>=1.24.0,<2
pillow>=10.0.0,<12
pytesseract>=0.3.10,<1
